+++ test/function/handle-calling-uncallable/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'does not give cryptic error when calling uncallable things (#1257)'
+};