+++ test/function/handle-calling-uncallable/main.js
@@ -1,0 +1,5 @@
+import foo from './foo.js';
+
+assert.throws( function () {
+	foo();
+}, /is not a function/ );