+++ test/function/this-not-namespace/main.js
@@ -1,0 +1,5 @@
+export class Foo {
+	constructor ( name ) {
+		this.name = undefined;
+	}
+}